Pruebas
Pruebas de Corrida, Medias y Varianza que devuelve si son aceptados en forma de String (Luego se puede cambiar). No se ha probado con datos aún xD.
Cambios Menores en Lehmer.

+++ Lehmer.java
@@ -2,15 +2,15 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Lehmer {
-    public static List<Long> listaNumLehmer = new ArrayList<>();
+    public static List<Integer> listaNumLehmer = new ArrayList<>();
 
-    public static long[] lehmer(int inicial, int n) {
-        long[] arrayNumeros = new long[n];
-        long aux, numAnterior;
+    public static Integer[] lehmer(int inicial, int n) {
+        Integer[] arrayNumeros = new Integer[n];
+        int aux, numAnterior;
         numAnterior = inicial;
         for (int y = 0; y < n; y++) {
             aux = (48271 * numAnterior) % ((int) Math.pow(2, 31) - 1);
-            arrayNumeros[y] = aux;
+            arrayNumeros[y] = (int) aux;
             numAnterior = aux;
         }
 
@@ -20,7 +20,7 @@ public class Lehmer {
     public static void main(String[] args) {
         int numInicial = 4808; // Semilla
         int n = 10; // Cantidad de números
-        long[] arrayLehmer = new long[n];
+        Integer[] arrayLehmer = new Integer[n];
         arrayLehmer = lehmer(numInicial, n);
 
         for (int x = 0; x < 50; x++) {
